% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dep.R
\docType{class}
\name{dep-class}
\alias{dep-class}
\title{An S4 class to represent differential expression and pvalues of samples/
probes between three groups.}
\description{
An S4 class to represent differential expression and pvalues of samples/
probes between three groups.
}
\section{Slots}{

\describe{
\item{\code{pvalues}}{A data frame containing the p-values, adjusted p-values,
and log2(fold changes) for all three comparisons between groups in the 
contrast factor, as well as optional multi-group tests.}

\item{\code{sampledata}}{Sample data with column ID and contrast}

\item{\code{contrast}}{The column name in `sampledata`` which contains the
three-group contrast factor used for comparisons.}

\item{\code{multi_group_test}}{An optional column name prefix for statistical tests 
between all three groups}

\item{\code{expression}}{An optional data frame or matrix containing the
expression data}
}}

