% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_volc3d.R
\name{DESeqToVolc}
\alias{DESeqToVolc}
\title{Convert DESeq2 objects to volcano3d}
\usage{
DESeqToVolc(
  object,
  objectLRT,
  contrast,
  data = NULL,
  pcutoff = 0.05,
  padj.method = "BH",
  ...
)
}
\description{
Convert DESeq2 objects to volcano3d
}
