% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Forest plot individual gene from 2x3 factor analysis}
\usage{
forest_plot(object, gene, scheme = c("red", "green3", "blue"), ...)
}
\arguments{
\item{object}{A 'volc3d' class object}

\item{gene}{Gene to plot}

\item{scheme}{Vector of 3 colours for plotting}

\item{...}{Optional arguments passed to [plot()]}
}
\description{
Forest plot individual gene from 2x3 factor analysis
}
