% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plotly}
\alias{forest_plotly}
\title{Forest plot individual gene from 2x3 factor analysis using plotly}
\usage{
forest_plotly(
  object,
  gene,
  scheme = c("red", "green3", "blue"),
  labs = NULL,
  error_type = c("ci", "se"),
  transpose = FALSE
)
}
\arguments{
\item{object}{A 'volc3d' class object from a 2x3 analysis generated by
\code{\link[=deseq_2x3_polar]{deseq_2x3_polar()}}}

\item{gene}{Gene to plot}

\item{scheme}{Vector of 3 colours for plotting}

\item{labs}{Optional character vector of labels for the groups}

\item{error_type}{Either "ci" or "se" to specify whether error bars use 95\%
confidence intervals or standard error.}

\item{transpose}{Logical whether to transpose the plot}
}
\description{
Forest plot individual gene from 2x3 factor analysis using plotly
}
\seealso{
\code{\link[=deseq_2x3_polar]{deseq_2x3_polar()}}
}
