% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_grid.R
\docType{class}
\name{grid-class}
\alias{grid-class}
\title{An S4 class to define the polar grid coordinates system.}
\description{
An S4 class to define the polar grid coordinates system.
}
\section{Slots}{

\describe{
\item{\code{polar_grid}}{The coordinates for the cylindrical grid segments with 
x,y,z coordinates}

\item{\code{axes}}{The axes features for plotly}

\item{\code{axis_labs}}{The axis labels}

\item{\code{r}}{The grid radius}

\item{\code{z}}{The grid height}

\item{\code{text_coords}}{data frame for axis label cartesian coordinates (x, y, z)}

\item{\code{n_r_breaks}}{The number of ticks on the r axis}

\item{\code{n_z_breaks}}{The number of ticks on the z axis}

\item{\code{r_breaks}}{The r axis ticks as a numeric}

\item{\code{z_breaks}}{The z axis ticks as a numeric}
}}

