% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_coords.R
\docType{class}
\name{polar-class}
\alias{polar-class}
\title{An S4 class to define the polar grid and coordinates for polar
differential expression plots}
\description{
An S4 class to define the polar grid and coordinates for polar
differential expression plots
}
\section{Slots}{

\describe{
\item{\code{sampledata}}{Sample data with column ID and contrast}

\item{\code{contrast}}{The column name in `sampledata`` which contains the
three-group contrast factor used for comparisons.}

\item{\code{multi_group_test}}{Column name prefix for statistical tests between
all three groups}

\item{\code{pvalues}}{A data frame containing the p-values, adjusted p-values,
and log2(fold changes) for all three comparisons between groups in the 
contrast factor, as well as optional multi-group tests.}

\item{\code{expression}}{An optional data frame or matrix containing the
expression data}

\item{\code{polar}}{A data.frame containing:
\itemize{
    \item{The mean expression for each of the three groups in comparison}
    \item{The z-score polar coordinates: 'y_zscore', 'x_zscore' and 
    'r_zscore'}
    \item{The fold-change polar coordinates: 'y_fc', 'x_fc' and 'r_fc'}
    \item{'angle': The angle in radians for polar coordinates}
    \item{'angle_degrees': The angle in degrees}
    \item{'maxExp': The maximally expressed group}
    \item{'sig': The significance group}
}}

\item{\code{non_sig_name}}{The category name for variables which are classed as not
significant}
}}

