% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_volc3d.R
\name{quick_volcano3d}
\alias{quick_volcano3d}
\title{Plot Quick 3d volcano plot}
\usage{
quick_volcano3d(obj, type = 1)
}
\arguments{
\item{obj}{Object of S3 class 'volc3d' following call to either
`polarCoords()` or `DESeqToVolc()`}

\item{type}{Either `1` or `2` specifying type of polar coordinates: `1` =
Z-scaled, `2` = unscaled (equivalent to log2 fold change for gene
expression).}
}
\description{
Slimline function for quick 3d volcano plot
}
