% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_coords_v2.R
\name{qval}
\alias{qval}
\title{Adjust P-values for Multiple Comparisons}
\usage{
qval(p, method = "qvalue")
}
\description{
Given a set of p-values, returns p-values adjusted using one of several
methods using methods available via [p.adjust] or [qvalue].
}
