% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{syn_example_meta}
\alias{syn_example_meta}
\title{PEAC synovial sample data}
\format{
A data frame with 81 rows and 4 variables:
\describe{
  \item{ID}{IDs which match column names in `syn_example_rld``}
  \item{Pathotype}{The sample pathotype}
}
}
\source{
\url{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1
}
}
\usage{
syn_example_meta
}
\description{
A dataset containing sample data for 81 synovial biopsies from the PEAC 
cohort
}
\keyword{datasets}
