% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{syn_example_rld}
\alias{syn_example_rld}
\title{PEAC synovial gene expression data}
\format{
A data frame with 100 rows representing the most significant 
genes/probes and 81 columns representing samples.
}
\source{
\url{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1
}
}
\usage{
syn_example_rld
}
\description{
A dataset containing the gene expression data for 81 synovial biopsies from 
the PEAC cohort
}
\keyword{datasets}
