% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_data.R
\docType{data}
\name{syn_metadata}
\alias{syn_metadata}
\title{PEAC synovial sample data}
\format{A data frame with 81 rows and 4 variables:
\describe{
  \item{ID}{IDs which match column names in `syn_rld``}
  \item{Batch}{weight of the diamond, in carats}
  \item{Gender}{Subject gender}
  \item{Pathotype}{The sample pathotype}
}}
\source{
\url{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1}
}
\usage{
syn_metadata
}
\description{
A dataset containing sample data for 81 synovial biopsies from the PEAC 
cohort
}
\keyword{datasets}
