% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Li_pvalues.R
\docType{data}
\name{syn_mod}
\alias{syn_mod}
\title{Li Module expression data in Synovium}
\format{A data frame with 346 rows representing modules and 81 columns 
representing samples}
\source{
\url{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1}
}
\usage{
syn_mod
}
\description{
Modular expression for synovial biopsies from the PEAC cohort
Li module expression data for synovial 81 samples
}
\details{
A dataset containing the module expression of Li modules among 81 synovial 
PEAC biopsies. Calculated using QuSAGE
}
\keyword{datasets}
