% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_data.R
\docType{data}
\name{syn_rld}
\alias{syn_rld}
\title{PEAC synovial gene expression data}
\format{A data frame with 16235 rows representing genes/probes and 81 
columns representing samples.}
\source{
\url{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1}
}
\usage{
syn_rld
}
\description{
A dataset containing the gene expression data for 81 synovial biopsies from 
the PEAC cohort
}
\keyword{datasets}
