% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_coords.R
\docType{class}
\name{volc3d-class}
\alias{volc3d-class}
\title{An S4 class to define the polar coordinates.}
\description{
An S4 class to define the polar coordinates.
}
\section{Slots}{

\describe{
\item{\code{df}}{List of coordinate data frames for scaled and unscaled expression}

\item{\code{outcome}}{Outcome vector}

\item{\code{data}}{Expression data}

\item{\code{pvals}}{Matrix or dataframe with p-values}

\item{\code{padj}}{Matrix adjusted p-values}

\item{\code{pcutoff}}{Cut-off for p-value significance}

\item{\code{scheme}}{Vector for colour scheme}

\item{\code{labs}}{Character vector for labelling groups}
}}

