% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano3D.R
\name{volcano3D}
\alias{volcano3D}
\title{Three-Dimensional Volcano Plot}
\usage{
volcano3D(
  polar,
  grid = NULL,
  fc_or_zscore = "zscore",
  label_rows = c(),
  label_size = 14,
  colour_scale = "discrete",
  axis_angle = 0.5,
  z_aspectratio = 1,
  xy_aspectratio = 1
)
}
\arguments{
\item{polar}{A polar object with created by \code{\link{polar_coords}}.}

\item{grid}{grid Optional grid list output by \code{\link{polar_grid3D}}. 
If NULL this will be calculated.}

\item{fc_or_zscore}{whether to use fold change or z-score for the p-values 
(default = 'zscore').}

\item{label_rows}{character vector of rows (names or indices) to label.}

\item{label_size}{font size for labels (default = 14).}

\item{colour_scale}{whether to use a "discrete" or "continuous" colour scale 
(default = "discrete").}

\item{axis_angle}{Angle in radians for the z axis (default = 0.5).}

\item{z_aspectratio}{The aspect ratio for the z axis compared to x and y 
(default = 1). Decreasing this makes the plot appear more squat.}

\item{xy_aspectratio}{The aspect ratio for the xy axis compared to z
(default = 1). Decreasing this makes the grid wider in the plot window.}
}
\value{
Returns a cylindrical 3D plotly plot featuring variables on a 
tri-axis radial graph with the -log10(multi-group test p-value) on the 
z-axis
}
\description{
Plots the pvalues from three-way comparisons in 3D space using plotly.
}
\examples{
library(volcano3Ddata)
data(syn_data)
syn_p_obj <- create_dep(sampledata = syn_metadata, 
                    contrast = "Pathotype", 
                    pvalues = syn_pvalues,
                    p_col_suffix = "pvalue", 
                    fc_col_suffix = "log2FoldChange",
                    multi_group_prefix = "LRT", 
                    expression = syn_rld)
syn_polar <- polar_coords(dep = syn_p_obj)
volcano3D(syn_polar, 
       label_rows = c("FMOD", "LAMP5", "TNNT3"), 
       label_size = 10, 
       xy_aspectratio = 1, 
       z_aspectratio = 0.9)
}
\references{
Lewis, Myles J., et al. (2019).
\href{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1}{
Molecular portraits of early rheumatoid arthritis identify clinical and 
treatment response phenotypes.}
\emph{Cell reports}, \strong{28}:9
}
\keyword{plot}
\keyword{polar,}
\keyword{pvalue,}
