% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano3D_v2.R
\name{volcano3dx}
\alias{volcano3dx}
\title{Three-Dimensional Volcano Plot}
\usage{
volcano3dx(
  obj,
  type = 1,
  grid_colour = "grey80",
  grid_width = 2,
  grid_options = NULL,
  axis_colour = "black",
  axis_width = 2,
  marker_size = 3,
  marker_outline_width = 0,
  marker_outline_colour = "white",
  z_axis_title_offset = 1.2,
  z_axis_title_size = 12,
  radial_axis_title_size = 14,
  radial_axis_title_offset = 1.2,
  axis_angle = 0.5,
  xy_aspectratio = 1,
  z_aspectratio = 0.8,
  camera_eye = list(x = 0.9, y = 0.9, z = 0.9),
  ...
)
}
\arguments{
\item{obj}{Object of S3 class 'volc3d' following call to either
`polarCoords()` or `DESeqToVolc()`}

\item{type}{Either `1` or `2` specifying type of polar coordinates: `1` =
Z-scaled, `2` = unscaled (equivalent to log2 fold change for gene
expression).}

\item{grid_colour}{The colour of the cylindrical grid (default "grey80")}

\item{grid_width}{The width of the grid lines (default 2)}

\item{grid_options}{Optional list of additional arguments to pass to
`polar_grid()`, eg. `z_axis_ticks` and `r_axis_ticks`}

\item{axis_colour}{The colour of the grid axes and labels (default "black")}

\item{axis_width}{The width of axis lines (default 2)}

\item{marker_size}{Size of the markers (default 3)}

\item{marker_outline_width}{Width for marker outline (default 0 means no
outline)}

\item{marker_outline_colour}{Colour for marker outline (default white)}

\item{axis_angle}{Angle in radians for the position of z axis (default 0.5)}

\item{xy_aspectratio}{The aspect ratio for the xy axis compared to z (default
0.75). Increasing this makes the grid wider in the plot window.}

\item{z_aspectratio}{The aspect ratio for the z axis compared to x and y.
Decreasing this makes the plot appear more squat. If `NULL` it is set
automatically.}

\item{camera_eye}{The (x,y,z) components of the start 'eye' camera vector.
This vector determines the view point about the origin of this scene.}

\item{...}{Optional arguments passed to `plot_ly`}
}
\description{
Version 2.0. Plots the three-way comparisons of variables such as gene
expression data in 3D space using plotly. x, y position represents polar
position on 3 axes representing the amount each variable or gene tends to
each of the 3 categories. The z axis represents -log10 P value for the
one-way test comparing each variable across the 3 groups.
}
