% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano4D.R
\name{volcano4D}
\alias{volcano4D}
\title{Add mode bar button to rotate the plot}
\usage{
volcano4D(
  p,
  rotate_icon_path = NULL,
  stop_icon_path = NULL,
  rotate_colour = "lightgreen",
  stop_colour = "tomato"
)
}
\arguments{
\item{p}{The volcano3D plot}

\item{rotate_icon_path}{The icon path for rotation}

\item{stop_icon_path}{The icon path for stop button}

\item{rotate_colour}{The colour for the rotate button (default='lightgreen')}

\item{stop_colour}{The colour for the stop button (default='tomato')}
}
\description{
Add mode bar button to rotate the plot
}
