% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_trio.R
\name{boxplot_trio}
\alias{boxplot_trio}
\title{Boxplot to compare three groups}
\usage{
boxplot_trio(
  dep,
  value,
  box_colours = c("green3", "blue", "red"),
  test = "t.test",
  levels_order = NULL,
  my_comparisons = NULL,
  text_size = 10,
  stat_size = 3,
  ...
)
}
\arguments{
\item{dep}{A dep object with the pvalues between groups of interest. Created
by \code{\link{create_dep}}.}

\item{value}{The row name or number in dep@expression to be analysed}

\item{box_colours}{The fill colours for each box 
(default = c('blue', 'red', 'green3') ).}

\item{test}{The statistical test used to compare expression. See method in
\code{\link[ggpubr]{stat_compare_means}}}

\item{levels_order}{A character vector stating the contrast groups to be 
plotted in order}

\item{my_comparisons}{A list of contrasts to pass to comparisons in 
\code{\link[ggpubr]{stat_compare_means}}. If NULL (default) all contrast 
pvalues are calculated and plotted.}

\item{text_size}{The font size of text (default = 10)}

\item{stat_size}{The font size of statistical parameter (default = 3).}

\item{...}{Other parameters for \code{\link[ggpubr]{stat_compare_means}}}
}
\value{
Returns a ggplot boxplot featuring the expression of one row in 
expression across the three groups in comparison with annotated pvalues.
}
\description{
Plots the expression of a specific row in expression comparing the three 
groups in a boxplot.
}
\examples{
data(syn_data)
syn_p_obj <- create_dep(sampledata = syn_metadata,
                    contrast = "Pathotype",
                    pvalues = syn_pvalues,
                    p_col_suffix ="pvalue",
                    fc_col_suffix = "log2FoldChange",
                    multi_group_prefix = "LRT",
                    expression = syn_rld)

boxplot_trio(syn_p_obj,
          value = "SLAMF6",
          test = "wilcox.test",
          levels_order = c("Lymphoid", "Myeloid", "Fibroid"),
          box_colours = c("blue", "red", "green3"))

boxplot_trio(syn_p_obj,
          value = "ITM2C",
          test = "wilcox.test",
          levels_order = c("Lymphoid", "Myeloid", "Fibroid"),
          my_comparisons = list(c("Lymphoid", "Myeloid"),
                                c("Myeloid", "Fibroid")),
          box_colours = c("blue", "red", "green3"))
}
\references{
Lewis, Myles J., et al. (2019). 
\href{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1}{
Molecular portraits of early rheumatoid arthritis identify clinical and 
treatment response phenotypes.} 
\emph{Cell reports}, \strong{28}:9
}
\keyword{boxplot}
\keyword{plot,}
\keyword{polar,}
\keyword{pvalue,}
