% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_coords_v2.R
\name{calc_pvals}
\alias{calc_pvals}
\title{Calculate one-way test and pairwise tests}
\usage{
calc_pvals(
  outcome,
  data,
  pcutoff = 0.05,
  padj.method = "BH",
  pairwise_test = c("t.test", "wilcoxon"),
  exact = FALSE
)
}
\description{
Internal function for calculating one-way test and pairwise tests (either
t-test or Wilcoxon test) on multi-column data against an outcome parameter
with 3 levels.
}
