% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_2x3_polar.R
\name{deseq_2x3_polar}
\alias{deseq_2x3_polar}
\title{Convert 2x3 design DESeq2 objects to a volcano3d object}
\usage{
deseq_2x3_polar(
  object,
  pcutoff = 0.05,
  padj.method = "BH",
  process = c("positive", "negative", "two.sided"),
  scheme = c("grey60", "red", "gold2", "green3", "cyan", "blue", "purple", "black"),
  labs = NULL
)
}
\arguments{
\item{object}{A list of 3 objects of class 'DESeqDataSet', or a list of 3
DESeq2 results dataframes generated by calling `DESeq2::results()`}

\item{pcutoff}{Cut-off for p-value significance}

\item{padj.method}{Can be any method available in `p.adjust` or `"qvalue"`.
The option `"none"` is a pass-through.}

\item{process}{Character value specifying colour process for statistical
significant genes: "positive" specifies genes are coloured if fold change
is >0; "negative" for genes with fold change <0 (note that for clarity the
polar position is altered so that genes along each axis have the most
strongly negative fold change values); or "two.sided" which is a compromise
in which positive genes are labelled as before but genes with negative fold
changes and significant p-values have an inverted colour scheme.}

\item{scheme}{Vector of colours starting with non-significant genes/variables}

\item{labs}{Optional character vector for labelling groups. Default `NULL`
leads to abbreviated labels based on levels in `outcome` using
[abbreviate()]. A vector of length 3 with custom abbreviated names for the
outcome levels can be supplied. Otherwise a vector length 8 is expected, of
the form "ns", "A+", "A+B+", "B+", "B+C+", "C+", "A+C+", "A+B+C+", where
"ns" means non-significant and A, B, C refer to levels 1, 2, 3 in
`outcome`, and must be in the correct order.}
}
\value{
Returns an S4 'volc3d' object containing:
\itemize{
  \item{'df'} A list of 2 dataframes. Each dataframe contains both x,y,z
  coordinates as well as polar coordinates r, angle. The first dataframe has
  coordinates based on the DESeq2 statistic. The 2nd dataframe is unscaled
  and represents log2 fold change for gene expression. The `type` argument in
  \code{\link{volcano3D}}, \code{\link{radial_plotly}} and
  \code{\link{radial_ggplot}} corresponds to these dataframes.
  \item{'outcome'} An empty factor whose levels are the three-group contrast
  factor for comparisons
  \item{'data'} Empty dataframe for compatibility
  \item{'pvals'} A dataframe containing p-values. Columns 1-3 are pairwise
  comparisons between the outcome factor for groups A, B, C respectively.
  \item{'padj'} A dataframe containing p-values adjusted for multiple testing
  \item{'pcutoff} Numeric value for cut-off for p-value significance
  \item{'scheme'} Character vector with colour scheme for plotting
  \item{'labs'} Character vector with labels for colour groups
}
}
\description{
This function is used instead of \code{\link{polar_coords}} if you have raw
RNA-Seq count data. It takes 3 `DESeqDataSet` objects, extracts statistical
results and converts the results to a 'volc3d' object, which can be directly
plotted.
}
