% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_coords.R
\name{polar_coords}
\alias{polar_coords}
\title{Coordinates for Three Way Polar Plot}
\usage{
polar_coords(
  dep,
  non_sig_name = "Not Significant",
  significance_cutoff = 0.01,
  fc_cutoff = 0.3
)
}
\arguments{
\item{dep}{A dep object with the pvalues between groups of interest. Created
by \code{\link{create_dep}}.}

\item{non_sig_name}{Name to assign non-significant points}

\item{significance_cutoff}{Value defining the significance cut-off (pvalues
below this will be coloured \code{non_sig_colour})}

\item{fc_cutoff}{The cut-off for fold change, below which markers will be
coloured according to `non_sig_colour`` (default = 0.3).}
}
\value{
Returns an S4 polar object containing:
\itemize{
  \item{'polar'} A data.frame containing:
  \itemize{
      \item{The mean expression for each of the three groups in comparison}
      \item{The z-score polar coordinates: 'y_zscore', 'x_zscore' and 
      'r_zscore'}
      \item{The fold-change polar coordinates: 'y_fc', 'x_fc' and 'r_fc'}
      \item{'angle': The angle in radians for polar coordinates}
      \item{'angle_degrees': The angle in degrees}
      \item{'hue': The continuous colour for volcano3D}
      \item{'col': The discrete colour for volcano3D}
      \item{'maxExp': The maximally expressed group}
      \item{'sig': The significance group}
  }
  \item{'pvalues'} A data frame containing the p-values, adjusted p-values,
  and log2(fold changes) for all three comparisons between groups in the 
  contrast factor, as well as optional multi-group tests.
  \item{'sampledata'} Sample data with column ID and contrast
  \item{'contrast'} The column name in `sampledata`` which contains the
  three-group contrast factor used for comparisons.
  \item{'multi_group_test'} Column name prefix for statistical tests between
  all three groups
  \item{'expression'} An optional data frame or matrix containing the
  expression data
  \item{'non_sig_name'} The category name for variables which are classed as 
  not significant
}
}
\description{
This function creates a data frame for downstream polar plots containing
the p-values from a three-way group comparison.
}
\examples{
library(volcano3Ddata)
data(syn_data)
syn_p_obj <- create_dep(sampledata = syn_metadata, 
                    contrast = "Pathotype", 
                    pvalues = syn_pvalues,
                    p_col_suffix = "pvalue", 
                    fc_col_suffix = "log2FoldChange",
                    multi_group_prefix = "LRT", 
                    expression = syn_rld)
syn_polar <- polar_coords(dep = syn_p_obj)
table(syn_polar@polar$sig) 
}
\references{
Lewis, Myles J., et al. (2019).
\href{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1}{
Molecular portraits of early rheumatoid arthritis identify clinical and
treatment response phenotypes.}
\emph{Cell reports}, \strong{28}:9
}
\keyword{plot}
\keyword{polar,}
\keyword{pvalue,}
