% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_grid.R
\name{polar_grid}
\alias{polar_grid}
\title{Polar Grid structure for Three Way Polar Plot}
\usage{
polar_grid(r_vector, axis_ticks, axis_angle = 4/3, ...)
}
\arguments{
\item{r_vector}{numerical vector of radial coordinates}

\item{axis_ticks}{a numerical vector of breaks for the radial axis}

\item{axis_angle}{Angle to position the radial axis in pi (default = 4/3)}

\item{...}{optional parameters for \link[base]{pretty}}
}
\value{
Returns a list containing:
\itemize{
  \item{'polar_grid'} The coordinates for a radial grid
  \item{'axes'} The axes features for plotly
  \item{'axis_labels'} The axis labels
  \item{'r'} The grid radius
  \item{'text_coords'} The coordinates for text labels
}
}
\description{
This function creates polar grid of radius r
}
\examples{
polar_grid(r_vector = 1:10, axis_ticks = 1:10)
}
\references{
Lewis, Myles J., et al. (2019).
\href{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1}{
"Molecular portraits of early rheumatoid arthritis identify clinical and
treatment response phenotypes.}
\emph{Cell reports}, \strong{28}:9
}
\keyword{plot}
\keyword{polar,}
\keyword{pvalue,}
