% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_ggplot.R
\name{radial_ggplot}
\alias{radial_ggplot}
\title{'Ggplot' for Three Way Polar Plot}
\usage{
radial_ggplot(
  polar,
  colours = NULL,
  non_sig_colour = "grey60",
  colour_scale = "discrete",
  continuous_shift = 120,
  label_rows = NULL,
  arrow_length = 1,
  grid = NULL,
  fc_or_zscore = "zscore",
  label_size = 5,
  axis_title_size = 5,
  axis_label_size = 3,
  marker_alpha = 0.7,
  marker_size = 3,
  axis_angle = 1/6,
  legend_size = 20,
  ...
)
}
\arguments{
\item{polar}{A polar object with the pvalues between groups of interest and 
polar coordinates. Created by \code{\link{polar_coords}}.}

\item{colours}{A named vector of colours for the groups. If NULL colours
will be assigned to c('green3', 'cyan', 'gold2', 'blue', 'purple' 'red'). If 
unnamed colours will be assigned in polar@polar$sig level order.}

\item{non_sig_colour}{non_sig_colour The colour for non-significant markers 
(default = "grey60").}

\item{colour_scale}{whether to use a "discrete" or "continuous" colour scale 
(default = "discrete").}

\item{continuous_shift}{the number of degrees (between 0 and 360) 
corresponding to the angle to offset the continuous colour scale by. The 
continuous colour scale is calculated by converting the angle to hue where 0 
degrees corresponds to red and 360 degrees to magenta (default = 120).}

\item{label_rows}{A vector of row names or numbers to label.}

\item{arrow_length}{length of label arrow (default = 1).}

\item{grid}{An optional grid object. If NULL this will be calculated using 
default values of  \code{\link{polar_grid}}.}

\item{fc_or_zscore}{Whether to use the z-score or fold change as magnitude.
Options are 'zscore' (default) or 'fc'.}

\item{label_size}{Font size of labels/annotations (default = 5).}

\item{axis_title_size}{Font size for axis titles (default = 5)}

\item{axis_label_size}{Font size for axis labels (default = 3)}

\item{marker_alpha}{The alpha parameter for markers (default = 0.7).}

\item{marker_size}{Size of the markers (default = 3).}

\item{axis_angle}{Angle for the radial axis labels in pi radians
(default = 1/6).}

\item{legend_size}{Size for the legend text (default = 20).}

\item{...}{Optional grid parameters to pass to 
\code{\link[volcano3D]{polar_grid}}.}
}
\value{
Returns a polar 'ggplot' object featuring variables on a tri-axis 
radial graph
}
\description{
This function creates a radar plot using 'ggplot' for a three-way comparison
}
\examples{
data(example_data)
syn_polar <- polar_coords(sampledata = syn_example_meta,
                          contrast = "Pathotype", 
                          groups = NULL, 
                          pvalues = syn_example_p, 
                          expression = syn_example_rld, 
                          p_col_suffix = "pvalue", 
                          padj_col_suffix = "padj", 
                          non_sig_name = "Not Significant", 
                          multi_group_prefix = "LRT",
                          significance_cutoff = 0.01, 
                          fc_cutoff = 0.3)
                          
col_vector = setNames(c('green', 'cyan', 'gold2', 'blue', 'purple'), 
                      c('Fibroid+', 'Fibroid+Lymphoid+', 
                        'Fibroid+Myeloid+', 'Lymphoid+', 
                        'Lymphoid+Myeloid+'))                             
                          
radial_ggplot(polar = syn_polar, label_rows = c("SLAMF6"), 
              colours=col_vector)
}
\references{
Lewis, Myles J., et al. (2019).
\href{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1}{
Molecular portraits of early rheumatoid arthritis identify clinical and
treatment response phenotypes.}
\emph{Cell reports}, \strong{28}:9
}
\keyword{hplot}
