% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_plotly.R
\name{radial_plotly}
\alias{radial_plotly}
\title{Three-way radial comparison Polar Plot (using 'plotly')}
\usage{
radial_plotly(
  polar,
  colours = NULL,
  non_sig_colour = "grey60",
  colour_scale = "discrete",
  continuous_shift = 120,
  label_rows = NULL,
  arrow_length = 50,
  grid = NULL,
  fc_or_zscore = "zscore",
  label_size = 14,
  axis_title_size = 16,
  axis_label_size = 10,
  axis_ticks = NULL,
  axis_angle = 5/6,
  ...
)
}
\arguments{
\item{polar}{A polar object with the pvalues between groups of interest and 
polar coordinates. Created by \code{\link{polar_coords}}.}

\item{colours}{A named vector of colours for the groups. If NULL colours
will be assigned to c('green3', 'cyan', 'gold2', 'blue', 'purple', 'red'). 
If unnamed colours will be assigned in polar@polar$sig level order.}

\item{non_sig_colour}{The colour for non-significant markers 
(default = "grey60").}

\item{colour_scale}{whether to use a 'discrete' or 'continuous' colour scale 
(default = 'discrete').}

\item{continuous_shift}{The number of degrees (between 0 and 360) 
to offset the continuous colour scale by. This is calculated by converting 
the angle to a hue using \code{\link[grDevices]{hsv}} where 0 corresponds to
the colour scale starting with red and 360 with magenta (default = 120).}

\item{label_rows}{A vector of row names or numbers to label.}

\item{arrow_length}{The length of label arrows (default = 50).}

\item{grid}{An optional grid object. If NULL this will be calculated using 
default values of  \code{\link{polar_grid}}.}

\item{fc_or_zscore}{Whether to use the z-score or fold change as magnitude.
Options are 'zscore' (default) or 'fc'.}

\item{label_size}{Font size of labels/annotations (default = 14)}

\item{axis_title_size}{Font size for axis titles (default = 16)}

\item{axis_label_size}{Font size for axis labels (default = 10)}

\item{axis_ticks}{A numerical vector of radial axis tick breaks. If
NULL this will be calculated using \code{\link[base]{pretty}}.}

\item{axis_angle}{Angle in  pi radians for the radial axis (default = 5/6).}

\item{...}{Optional grid parameters to pass to 
\code{\link[volcano3D]{polar_grid}}.}
}
\value{
Returns a 'plotly' plot featuring variables on a tri-axis
radial graph
}
\description{
This function creates an interactive 'plotly' object which maps differential 
expression onto a polar coordinates.
}
\examples{
data(example_data)
syn_polar <- polar_coords(sampledata = syn_example_meta,
                          contrast = "Pathotype", 
                          groups = NULL, 
                          pvalues = syn_example_p, 
                          expression = syn_example_rld, 
                          p_col_suffix = "pvalue", 
                          padj_col_suffix = "padj", 
                          non_sig_name = "Not Significant", 
                          multi_group_prefix = "LRT",
                          significance_cutoff = 0.01, 
                          fc_cutoff = 0.3)

col_vector = setNames(c('green', 'cyan', 'gold2', 'blue', 'purple'), 
                      c('Fibroid+', 'Fibroid+Lymphoid+', 'Fibroid+Myeloid+', 
                        'Lymphoid+', 'Lymphoid+Myeloid+'))      
                                                                        
radial_plotly(polar = syn_polar, label_rows = c("SLAMF6"), 
              colours=col_vector)
}
\references{
Lewis, Myles J., et al. (2019).
\href{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1}{
Molecular portraits of early rheumatoid arthritis identify clinical and
treatment response phenotypes.}
\emph{Cell reports}, \strong{28}:9
}
\keyword{hplot}
\keyword{iplot}
