% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_plotly.R
\name{radial_plotly}
\alias{radial_plotly}
\title{Three-way radial comparison Polar Plot (using plotly)}
\usage{
radial_plotly(polar, fc_cutoff = 0.3, non_sig_colour = "grey60",
  label_rows = NULL, grid = NULL, fc_or_zscore = "zscore",
  label_size = 14, axis_title_size = 16, axis_label_size = 10,
  arrow_length = 50, axis_ticks = NULL, axis_angle = 5/6, ...)
}
\arguments{
\item{polar}{The coordinates for plotting onto a polar plot}

\item{fc_cutoff}{The cut-off for fold change, below which markers will be
coloured according to `non_sig_colour`` (default = 0.3).}

\item{non_sig_colour}{The colour for non-significant markers according to 
fold change.}

\item{label_rows}{A character vector of genes or row numbers to label}

\item{grid}{Optional grid list output by \code{\link{polar_grid}}. If NULL 
this will be calculated.}

\item{fc_or_zscore}{Whether to use the z-score or fold change as magnitude
(options are c("zscore", "fc"))}

\item{label_size}{Font size of labels (default = 14)}

\item{axis_title_size}{Font size for axis titles (default = 16)}

\item{axis_label_size}{Font size for axis labels (default = 10)}

\item{arrow_length}{length of label arrow (default = 50)}

\item{axis_ticks}{a numerical vector of tick breaks for the radial axis. If
NULL this will be calculated using \code{\link[base]{pretty}}.}

\item{axis_angle}{Angle for the axis labels (default = 5/6).}

\item{...}{Optional parameters to pass to \code{\link[base]{pretty}} to 
define the radial axis ticks.}
}
\value{
Returns a polar plotly plot featuring variables on a tri-axis
radial graph
}
\description{
This function creates a plot using plotly which maps the expression
onto a polar coordinates.
}
\examples{
library(volcano3Ddata)
data(syn_data)
syn_p_obj <- create_dep(sampledata = syn_metadata, 
                    contrast = "Pathotype", 
                    pvalues = syn_pvalues,
                    p_col_suffix = "pvalue", 
                    fc_col_suffix = "log2FoldChange",
                    multi_group_prefix = "LRT", 
                    expression = syn_rld)
syn_polar <- polar_coords(dep = syn_p_obj)
radial_plotly(polar=syn_polar, 
            fc_cutoff=0.1, 
            label_rows=c("SLAMF6", "PARP16", "ITM2C"))
}
\references{
Lewis, Myles J., et al. (2019).
\href{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1}{
Molecular portraits of early rheumatoid arthritis identify clinical and
treatment response phenotypes.}
\emph{Cell reports}, \strong{28}:9
}
\keyword{plot,}
\keyword{plotly,}
\keyword{polar,}
\keyword{pvalues,}
\keyword{radial}
