% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{syn_example_meta}
\alias{syn_example_meta}
\title{PEAC synovial sample data}
\format{
A data frame with 81 rows and 1 variables:
\describe{
\item{Pathotype}{The synovial biopsy histological pathotype}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/31461658/
}
}
\usage{
syn_example_meta
}
\description{
A dataset containing sample data for 81 synovial biopsies from the PEAC
cohort
}
\keyword{datasets}
