% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{syn_example_p}
\alias{syn_example_p}
\title{Synovial differential expression of genes across pathotypes in PEAC cohort}
\format{
A data frame with 100 rows representing the most significant 
genes/probes and 13 columns for each statistical parameter:
\describe{
    \item{Module}{The gene name}
    \item{Fibroid-Lymphoid pvalue}{pvalue from fibroid vs lymphoid 
    comparison}
    \item{Fibroid-Lymphoid padj}{adjusted pvalue from fibroid vs lymphoid 
    comparison}
    \item{Fibroid-Lymphoid log2FoldChange}{logarithmic fold change from 
    fibroid vs lymphoid comparison}
    \item{Lymphoid-Myeloid pvalue}{pvalue from lymphoid vs myeloid 
    comparison}
    \item{Lymphoid-Myeloid padj}{adjusted pvalue from lymphoid vs myeloid 
    comparison}
    \item{Lymphoid-Myeloid log2FoldChange}{logarithmic fold change from 
    lymphoid vs myeloid comparison}
    \item{Myeloid-Fibroid pvalue}{pvalue from myeloid vs fibroid comparison}
    \item{Myeloid-Fibroid padj}{adjusted pvalue from myeloid vs fibroid 
    comparison}
    \item{Myeloid-Fibroid log2FoldChange}{logarithmic fold change from 
    myeloid vs fibroid comparison}
    \item{LRT pvalue}{pvalue from three-way likelihood ratio comparison}
    \item{LRT padj}{adjusted pvalue from three-way likelihood ratio 
    comparison}
}
}
\source{
\url{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1
}
}
\usage{
syn_example_p
}
\description{
A dataset containing the differential expression parameters between different
pathotype groups for 81 synovial biopsies from the PEAC cohort.
}
\keyword{datasets}
