% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano3D.R
\name{volcano3D}
\alias{volcano3D}
\title{Three-Dimensional Volcano Plot}
\usage{
volcano3D(
  polar,
  colours = NULL,
  non_sig_colour = "grey60",
  colour_scale = "discrete",
  continuous_shift = 120,
  label_rows = c(),
  grid = NULL,
  fc_or_zscore = "zscore",
  label_size = 14,
  axis_angle = 0.5,
  z_aspectratio = 1,
  xy_aspectratio = 1,
  ...
)
}
\arguments{
\item{polar}{A polar object with created by \code{\link{polar_coords}}.}

\item{colours}{A named vector of colours for the groups. If NULL colours
will be assigned to c('green3', 'cyan', 'gold2', 'blue', 'purple' 'red'). If 
unnamed colours will be assigned in polar@polar$sig level order.}

\item{non_sig_colour}{The colour for non-significant markers according to 
fold change (default='grey60').}

\item{colour_scale}{whether to use a 'discrete' or 'continuous' colour scale 
(default = 'discrete').}

\item{continuous_shift}{the number of degrees (between 0 and 360) 
corresponding to the angle to offset the continuous colour scale by. The 
continuous colour scale is calculated by converting the angle to hue where}

\item{label_rows}{A vector of row names or numbers to label.}

\item{grid}{An optional grid object. If NULL this will be calculated using 
default values of  \code{\link{polar_grid}}.}

\item{fc_or_zscore}{whether to use fold change or z-score for the p-values 
(default = 'zscore').}

\item{label_size}{font size for labels (default = 14).}

\item{axis_angle}{Angle in radians for the z axis (default = 0.5).}

\item{z_aspectratio}{The aspect ratio for the z axis compared to x and y 
(default = 1). Decreasing this makes the plot appear more squat.}

\item{xy_aspectratio}{The aspect ratio for the xy axis compared to z
(default = 1). Decreasing this makes the grid wider in the plot window.}

\item{...}{Optional parameters to pass to \code{\link{polar_grid}}.}
}
\value{
Returns a cylindrical 3D plotly plot featuring variables on a 
tri-axis radial graph with the -log10(multi-group test p-value) on the 
z-axis
}
\description{
Plots the pvalues from three-way comparisons in 3D space using plotly.
}
\examples{
\dontrun{
library(volcano3Ddata)
data(syn_data)
syn_polar <- polar_coords(sampledata = syn_metadata,
                       contrast = "Pathotype",
                       pvalues = syn_pvalues, 
                       expression = syn_rld, 
                       p_col_suffix = "pvalue", 
                       padj_col_suffix = "padj", 
                       non_sig_name = "Not Significant", 
                       multi_group_prefix = "LRT",
                       significance_cutoff = 0.01, 
                       fc_cutoff = 0.3)

volcano3D(syn_polar, 
    label_rows = c("FMOD", "LAMP5", "TNNT3"), 
    label_size = 10, 
    xy_aspectratio = 1, 
    z_aspectratio = 0.9)
}
}
\references{
Lewis, Myles J., et al. (2019).
\href{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1}{
Molecular portraits of early rheumatoid arthritis identify clinical and 
treatment response phenotypes.}
\emph{Cell reports}, \strong{28}:9
}
\keyword{plot}
\keyword{polar,}
\keyword{pvalue,}
